fix symbian build, fix few bugs
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@1729 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/engine/types.h
@@ -109,8 +109,8 @@ void WriteLE16(u8 *p, u16 x);
 namespace std
 {
     int c_abs(int x);
-    int c_abs(float x);
-    int c_abs(double x);
+    float c_abs(float x);
+    double c_abs(double x);
     int c_isspace(char c);
 
 #define isspace(c) c_isspace(c)
